% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice.tree.R
\name{choice.tree}
\alias{choice.tree}
\title{Optimal subtree choice}
\usage{
choice.tree(seq, coeff = 0.1)
}
\arguments{
\item{seq}{sequence of subtrees pruned from a maximal tree. Currently provided by functions prune.tree from package tree, spatprune and spatcart.}

\item{coeff}{a scalar parameter adjusting the size of the modified maximal gap slope heuristic. Default is to take 0.1x(size of the largest plateau).}
}
\value{
a tibble object with 2 arguments
}
\description{
Optimal subtree choice using slope heuristics proposed in "Spatial Classification Trees" by Bar-Hen, Gey, Poggi.
}
\examples{
library(tree)
t = tree(Species~., data = iris)
seq = prune.misclass(t)
choice.tree(seq)

}
\references{
\emph{Spatial Classification Trees}, by A. Bar-Hen, S. Gey and J.-M. Poggi (2020).
}
