% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.partition.tree.R
\name{gg.partition.tree}
\alias{gg.partition.tree}
\title{Partition of a tree (requires ggplot2)}
\usage{
gg.partition.tree(tree, ordvars, ...)
}
\arguments{
\item{tree}{an object of class "tree". Currently provided by the functions spatcart or spattree.}

\item{ordvars}{the ordering of the variables to be used in a 2D plot. Specify the names in a character string of length 2; the first will be used on the x axis.}

\item{...}{graphical parameters.}
}
\value{
an object of class "ggplot".
}
\description{
Plots the partition of a tree involving one or two variables. Requires package ggplot2.
}
\examples{
library(tree)
tmax = tree(Species~Sepal.Length+Petal.Length, data = iris)
t = prune.misclass(tmax, best = 3)

library(dplyr)
library(ggplot2)
iris \%>\% select(Sepal.Length, Petal.Length, Species) \%>\%
ggplot(aes(Sepal.Length, Petal.Length, color = Species))+geom_point()+
gg.partition.tree(t)
}
