% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatcart.innerfunctions.R
\name{impurity}
\alias{impurity}
\title{spatcart inner function}
\usage{
impurity(y, r, numvar, wt, x, Kt, parms)
}
\arguments{
\item{y}{spatial bivariate marked point process of class 'ppp'.}

\item{r}{scale resolution parameter.}

\item{numvar}{index of variable used.}

\item{wt}{inner parameter.}

\item{x}{ordered variable numvar.}

\item{Kt}{parent node impurity.}

\item{parms}{inner parameters.}
}
\value{
a data.frame object giving impurity loss at each scale value for each split along x ordered values.
}
\description{
Impurity loss function on ordered variable x, at scales between 0 and r, starting form parent nodes's impurity Kt.
}
