% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatcart.innerfunctions.R
\name{make.grid}
\alias{make.grid}
\title{spatcart inner function}
\usage{
make.grid(tree, ypp, K, d = 100)
}
\arguments{
\item{tree}{an object of class "tree". Currently provided by the functions spatcart or spattree.}

\item{ypp}{the spatial bivaraitte marked point process used to generate tree a, of class 'ppp' (see spatstat).}

\item{K}{the values of interactions between marks of the point process ypp restricted to the leaves of tree a. Interactions are computed through the Ripley's Kcross function. They can be found in output values of functions spatcart or spattree.}

\item{d}{a scalar parameter defining the dxd grid resolution to produce the heatmap associated to the leaves of tree a and values of K. Default value is d = 100, leading to a 10e+4 grid of points.}
}
\value{
a list with 2 objects : the grid data.frame and the mesh = d resolution
}
\description{
Computation of the expanded heatmap grid associated to the leaves of a tree and the interactions between marks of a bivariate marked point process.
}
