% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paracou.R
\name{Paracou}
\alias{Paracou}
\title{CART and SpatCART on the Paracou data set}
\usage{
Paracou(r, minsplit = 10, minleaf = 5, graph = TRUE)
}
\arguments{
\item{r}{the initial resolution used by SpatCART to construct the maximal tree (see spattree).}

\item{minsplit}{minimal size for the node to be split. Default is minsplit = 10 and.}

\item{minleaf}{minimal size of each resulting leaf for a node to be split. Default is minleaf = 5.}

\item{graph}{logical. If TRUE (Default), also produce the graphs of the number of leaves of the pruned subtrees sequence with respect to the complexity parameter used in the pruning algorithm. The triangle symbol represents the tree selected via the modified largest jump method slpe heuristics, while the diamond symbol represents the tree selected via the largest plateau method one.}
}
\value{
a list of 7 objects of class "ggplot".
}
\description{
Produce graphical results of CART and SpatCART algorithms at fixed initial resolution for the Paracou data set.
}
\examples{
Paracou(15)
}
\references{
\emph{Spatial Classification Trees}, by A. Bar-Hen, S. Gey and J.-M. Poggi (2021).
}
