% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.spattree.R
\name{partition.spattree}
\alias{partition.spattree}
\title{Heatmap of interaction between marks}
\usage{
partition.spattree(a, ypp, K, d = 100)
}
\arguments{
\item{a}{an object of class "tree". Currently provided by the functions spatcart or spattree.}

\item{ypp}{the bivariate marked spatial point process used to generate tree a, of class 'ppp' (see spatstat).}

\item{K}{the values of interactions between marks of the point process ypp restricted to the leaves of tree a. Interactions are computed through the Ripley's Kcross function. They can be found in output values of functions spatcart or spattree.}

\item{d}{a scalar parameter defining the dxd grid resolution to produce the heatmap associated to the leaves of tree a and values of K. Default value is d = 100, leading to a 10e+4 grid of points.}
}
\value{
an object of class "ggplot".
}
\description{
Plot the heatmap induced by a tree partition of interaction between marks for a bivariate marked spatial point process
}
\examples{
library(spatstat)
library(spatcart)
# Loading Paracou data set
ypp = Paracoudata()
r = 15
t = spatcart(ypp, r, graph = FALSE)

a = t$opt.tree.min
K = t$K[rownames(a$frame)][a$frame$var=="<leaf>"]

partition.spattree(a, ypp, K)
}
\references{
\emph{Spatial Classification Trees}, by A. Bar-Hen, S. Gey and J.-M. Poggi (2021).
}
