% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.tree.corrected.R
\name{partition.tree.corrected}
\alias{partition.tree.corrected}
\title{Partition of a tree}
\usage{
partition.tree.corrected(tree, label = "yval", add = FALSE, ordvars, ...)
}
\arguments{
\item{tree}{an object of class "tree". Currently provided by the functions spatcart or spattree.}

\item{label}{a character string giving the column of the frame component of tree to be used to label the regions.}

\item{add}{if true, add to existing plot, otherwise start a new plot.}

\item{ordvars}{the ordering of the variables to be used in a 2D plot. Specify the names in a character string of length 2; the first will be used on the x axis.}

\item{...}{graphical parameters.}
}
\value{
None
}
\description{
Plot the partition of a tree involving one or two variables. Corrected version of the {tree} function partition.tree to be able to handle trees partitioning on only one variable.
}
