% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repulsion.R
\name{repulsion}
\alias{repulsion}
\title{Repulsive simulated data set}
\usage{
repulsion(n, r, h = 0.4, ncase = 2, graph = TRUE)
}
\arguments{
\item{n}{intensity of the Poisson point process used to simulate data (see spatstat).}

\item{r}{value of the local repulsive scale. 0< r <1.}

\item{h}{margin parameter defining the mixing proportion of marks inside window. Default is h = 0.4.}

\item{ncase}{number of marks of the marked point process. Default is ncase = 2.}

\item{graph}{logical. If TRUE (Default), plot the data scatter plot.}
}
\value{
a data.frame object with variables (x, y, marks).
}
\description{
Simulation of a locally repulsive spatial marked point process
}
\examples{
repulsion(1000, 0.05)
repulsion(1000, 0.1, h = 0.1)
}
