% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spattree.R
\name{spattree}
\alias{spattree}
\title{Spatial Classification Maximal Tree}
\usage{
spattree(
  y,
  r,
  ties = TRUE,
  offset = NULL,
  wt = NULL,
  parms = NULL,
  minsplit = 10,
  minleaf = 5,
  mindev = 0
)
}
\arguments{
\item{y}{spatial bivariate marked point process of class 'ppp' (see spatstat).}

\item{r}{initial resolution scale to use to construct spatcart maximal tree. Decreasing with splitting..}

\item{ties}{logical. If TRUE (default), in case of ties in splitting rule, the first split achieving maximum loss in impurity is taken. If FALSE, scale adaptation inside nodes is made to avoid ties.}

\item{offset}{parameter related to the tree() function of package tree.}

\item{wt}{vector of non-negative observational weights; fractional weights are allowed.}

\item{parms}{other parameters related to the tree function of package tree.}

\item{minsplit}{minimal size for the node to be split. Default is minsplit = 10.}

\item{minleaf}{minimal size of each resulting leaf for a node to be split. Default is minleaf = 5.}

\item{mindev}{minimal value of node impurety to control node splitting. Default is mindev = 0.}
}
\value{
a list of objects containing the maximal tree; the sequence of pruned subtrees and the corresponding complexities; the values of scale resolution and impurity inside nodes of the maximal tree.
}
\description{
Spatial Classification Maximal Tree construction using the Ripley's intertype K-function to partition spatial bivariate marked point process. Requires packages "spatstat" and "tree".
}
\details{
parameter minsplit must be at least 2xminleaf.
}
\examples{
library(spatcart)
# Loading Paracou data set
ypp = Paracoudata()

# Keeping ties in splitting rules
t = spattree(ypp, 15)

# optimizing ties in splitting rules
t = spattree(ypp, 15, ties = FALSE)

# Changing stopping rule for splitting
t = spattree(ypp, 15, minsplit = 20, minleaf = 10)
}
\references{
\emph{Spatial Classification Trees}, by A. Bar-Hen, S. Gey and J.-M. Poggi, HAL 01837065 (2018).
}
