% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatcart.innerfunctions.R
\name{spetemp}
\alias{spetemp}
\title{spatcart inner function}
\usage{
spetemp(y, wt, parms)
}
\arguments{
\item{y}{bivariate marked point process of class 'ppp'.}

\item{wt}{inner parameter}

\item{parms}{inner parameter}
}
\value{
node evaluation
}
\description{
Node evaluation
}
