% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatcart.innerfunctions.R
\name{spfit}
\alias{spfit}
\title{spatcart inner function}
\usage{
spfit(
  ynode,
  r,
  ties,
  offset,
  wt,
  parms,
  sub,
  nodeindex,
  minsplit,
  minleaf,
  mindev
)
}
\arguments{
\item{ynode}{spatial bivariaye marked point process restricted to node, of class 'ppp'.}

\item{r}{scale resolution at node}

\item{ties}{logical. If TRUE, in case of ties in splitting rule, the first split achieving maximum loss in impurity is taken. If FALSE, scale adaptation inside nodes is made to avoid ties.}

\item{offset}{inner parameter.}

\item{wt}{inner parameter.}

\item{parms}{inner parameters.}

\item{sub}{subset indexes of observations in node.}

\item{nodeindex}{node index.}

\item{minsplit}{minimal size of each resulting leaf for a node to be split.}

\item{minleaf}{minimal size of each resulting leaf for a node to be split.}

\item{mindev}{minimal value of node impurety to control node splitting.}
}
\value{
a list of objects : the maximal tree of class 'tree', and 2 vectors giving inside nodes resolution scales and impurity.
}
\description{
Maximal tree computation. Produces list of "frame" and "where" to put in class "tree".
}
