% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatcart.innerfunctions.R
\name{spitemp}
\alias{spitemp}
\title{spatcart inner function}
\usage{
spitemp(y, offset, parms, wt)
}
\arguments{
\item{y}{bivariate marked point process of class 'ppp'.}

\item{offset}{inner parameter}

\item{parms}{inner parameter}

\item{wt}{inner parameter}
}
\value{
node initialisation.
}
\description{
Node initialisation.
}
