% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatcart.innerfunctions.R
\name{spstemp}
\alias{spstemp}
\title{spatcart inner function}
\usage{
spstemp(y, r, numvar, wt, x, Kt, parms)
}
\arguments{
\item{y}{spatial bivariate marked point process of class 'ppp'.}

\item{r}{scale resolution}

\item{numvar}{index of splitting variable}

\item{wt}{inner parameter.}

\item{x}{orderd values of splitting variable numvar}

\item{Kt}{parent node impurity.}

\item{parms}{inner parameters.}
}
\value{
a list with the goodness and imputity losses at each split along ordered variable x.
}
\description{
Goodness of splitting rules on ordered variable x
}
