% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paracoudata.R
\name{Paracoudata}
\alias{Paracoudata}
\title{Load Paracou data set}
\usage{
Paracoudata()
}
\value{
the bivaraite spatial point process of class 'ppp'.
}
\description{
Load the Paracou data set as a bivaraite spatial point process.
}
\examples{
Paracoudata()
}
